+++ app/src/main/java/uz/texnopos/elektrolife/data/model/newsale/OrderItem.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 data class OrderItem(
     @SerializedName("product_id")
     var productId: Int,
-    var count: Int,
+    var count: Double,
     @SerializedName("unit_id")
     val unitId: Int,
     var price: Double
